Bug to enable http://localhost:8080/analyzecsv.html fixed in SimulationEventManager.

+++ scratch/tcpfiles/NodeStatistics.h
@@ -108,8 +108,8 @@ public:
     Time FirmwareTransferTime;
 
     Time TimeStreamStarted = Time(0);
-    long IPCameraTotalDataSent = 0;
-    long IPCameraTotalDataReceivedAtAP = 0;
+    double IPCameraTotalDataSent = 0;
+    double IPCameraTotalDataReceivedAtAP = 0;
     Time IPCameraTotalTimeSent = Time(0);
 
     double getIPCameraSendingRate();
